+++ changes.md
@@ -5,6 +5,7 @@
 * Added WP-CLI `wp retsci` commands.
 * Added `wrc::manage_property::before_update` action to XMLRPC class.
 * Added [/wp-json/wp-rets-client/v1/cleanupProcess] endpoint. It recalculates terms counts.
+* Deleting and trashing of property requires rets_id (MLS ID) instead of post ID.
 * Cleaned up XMLRPC class.
 
 #### 0.3.6